requirements.txt
django==3.0
djangorestframework==3.12.4
django-mssql-backend==2.8.1
django-cors-headers==3.10.0
boto3
celery===5.1.2 
gevent==20.9.0
django-celery-beat==2.2.1
pyodbc